saudi-ai-audit/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dateutil==2.8.2
hijri-converter==2.3.1
arabic-reshaper==3.0.0
python-bidi==0.4.2
reportlab==4.0.7
openpyxl==3.1.2
scipy==1.11.3
rich==13.7.0
pytest==7.4.3